+++ flutter_inappwebview_forge_windows/lib/src/main.dart
@@ -11,3 +11,4 @@ export 'webview_environment/main.dart';
 export 'web_message/main.dart';
 export 'web_notification/main.dart';
 export 'container_controller.dart';
+export 'pull_to_refresh_controller.dart';